refactor: remove Executor abstraction
The executor abstraction is not necessary as TDP is build on top of Ansible.
