+++ tdp/core/deployment/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2022 TOSIT.IO
 # SPDX-License-Identifier: Apache-2.0
 
-from .ansible_executor import AnsibleExecutor
 from .deployment_iterator import DeploymentIterator
 from .deployment_plan import (
     DeploymentPlan,
